Move Subtitle component into components directory
At top-level by mistake
